BuildResume/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.34.18
botocore==1.34.18
chardet==5.2.0
Django==4.2.9
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.2.1
pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.0.9
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==1.26.18
